Migrate database and services from SQLite to PostgreSQL
Replaced SQLite with PostgreSQL across services, models, and database operations. Updated schema, migrations, and queries for PostgreSQL compatibility. Adjusted types and added `chrono` for timestamp handling.

+++ migrations/20250522000000_add_duration_column.sql
@@ -1,7 +1,7 @@
 -- Add duration column to videos table
 -- Up migration
 
-ALTER TABLE videos ADD COLUMN duration INTEGER;
+ALTER TABLE videos ADD COLUMN duration BIGINT;
 
 -- Down migration
 -- ALTER TABLE videos DROP COLUMN duration;
